+++ site-cookbooks/alterf/attributes/default.rb
@@ -8,3 +8,7 @@ default[:alterf][:shared_dirs] = %w[log bundle tmp/files]
 default[:alterf][:requirements] = %w[mysql-devel]
 default[:alterf][:ruby_version] = '2.5.5'
 default[:alterf][:mysql_users] = %w[development test]
+default[:alterf][:python] = {
+  version: '3.6.6',
+  packages: %w[mysql-connector-python-rf pyyaml pandas scikit-learn],
+}