+++ site-cookbooks/pyenv/metadata.rb
@@ -1,7 +1,6 @@
-name             'sphinx'
+name             'pyenv'
 maintainer       'Leonis0813'
 license          'All rights reserved'
-description      'Installs/Configures sphinx'
+description      'Installs/Configures pyenv'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
-depends          'pyenv'